mrc-misc-fn-modules(extract-chunked-prefix-gate-move,mechanical_provable): move maybe_disable_chunked_prefix_cache to misc_utils module (cut+paste)
